Migrate GPI/Checkers tests
